refactor(core.utils): split utilities into multiple files
